Add browser support to benchmark suite
