+++ package.json
@@ -30,11 +30,12 @@
     "cover": "istanbul cover -x out/*.js -- test/index.js",
     "check-prettier": "yarn prettier && git diff --exit-code",
     "check-docs": "yarn build-docs && git diff --exit-code",
-    "benchmarks-es6": "node test/benchmarks.js",
     "dist": "yarn dist-clean && mkdir -p dist/out && buble -i lib -o dist/lib && cp out/libh3.js dist/out",
     "dist-clean": "rm -rf dist",
     "dist-test": "yarn dist && buble -i test -o dist/test",
-    "benchmarks": "yarn dist-test && node dist/test/benchmarks.js",
+    "dist-benchmark": "yarn dist && buble -i benchmark -o dist/benchmark",
+    "benchmark-node": "yarn dist-benchmark && node dist/benchmark/node.js",
+    "benchmark-browser": "yarn dist-benchmark && budo dist/benchmark/browser.js -t babelify --open --title 'h3-js benchmarks'",
     "prepublish": "yarn dist",
     "init-docker": "docker run -dit --name emscripten -v $(pwd):/src trzeci/emscripten:sdk-tag-1.37.40-64bit bash",
     "build-emscripten": "docker exec -it emscripten bash .build-emscripten.sh",
@@ -44,6 +45,7 @@
   "devDependencies": {
     "benchmark": "^2.1.4",
     "buble": "^0.19.3",
+    "budo": "^11.5.0",
     "eslint": "^4.19.1",
     "eslint-config-prettier": "^2.9.0",
     "eslint-config-uber-es2015": "^3.1.2",
